Split a lot of logic out into AnomalyScannerSystem.
Re-use bar drawing buffer.

Clean up and comment the generation of the dynamic texture.

GetScannerMessage is in AnomalySystem because it's just getting information about the anomaly. We could re-use this to have the same information be displayed when the vessel is used as well.
