+++ Content.Client/Anomaly/AnomalySystem.cs
@@ -31,8 +31,6 @@ public sealed partial class AnomalySystem : SharedAnomalySystem
         SubscribeLocalEvent<AnomalyComponent, AnimationCompletedEvent>(OnAnimationComplete);
 
         SubscribeLocalEvent<AnomalySupercriticalComponent, ComponentShutdown>(OnShutdown);
-
-        InitializeScanner();
     }
 
     private void OnStartup(EntityUid uid, AnomalyComponent component, ComponentStartup args)
