+++ Content.Server/Anomaly/Components/SecretDataAnomalyComponent.cs
@@ -10,7 +10,7 @@ namespace Content.Server.Anomaly.Components;
 /// <summary>
 /// Hides some information about the anomaly when scanning it
 /// </summary>
-[RegisterComponent, Access(typeof(SecretDataAnomalySystem), typeof(AnomalySystem))]
+[RegisterComponent, Access(typeof(SecretDataAnomalySystem), typeof(AnomalySystem), typeof(AnomalyScannerSystem))]
 public sealed partial class SecretDataAnomalyComponent : Component
 {
     /// <summary>
