+++ Content.Shared/Anomaly/Components/AnomalyScannerComponent.cs
@@ -16,7 +16,7 @@ namespace Content.Shared.Anomaly.Components;
 /// This is used for scanning anomalies and
 /// displaying information about them in the ui
 /// </summary>
-[RegisterComponent, Access(typeof(SharedAnomalySystem))]
+[RegisterComponent, Access(typeof(SharedAnomalyScannerSystem))]
 [NetworkedComponent]
 public sealed partial class AnomalyScannerComponent : Component
 {
@@ -29,12 +29,12 @@ public sealed partial class AnomalyScannerComponent : Component
     /// <summary>
     /// How long the scan takes
     /// </summary>
-    [DataField("scanDoAfterDuration")]
+    [DataField]
     public float ScanDoAfterDuration = 5;
 
     /// <summary>
     /// The sound plays when the scan finished
     /// </summary>
-    [DataField("completeSound")]
+    [DataField]
     public SoundSpecifier? CompleteSound = new SoundPathSpecifier("/Audio/Items/beep.ogg");
 }
