+++ Content.Shared/Anomaly/SharedAnomalyScannerSystem.cs
@@ -1,0 +1,13 @@
+﻿using Content.Shared.Popups;
+using Robust.Shared.Audio.Systems;
+using Robust.Shared.Timing;
+
+namespace Content.Shared.Anomaly;
+
+public abstract class SharedAnomalyScannerSystem : EntitySystem
+{
+    [Dependency] protected readonly SharedPopupSystem Popup = default!;
+    [Dependency] protected readonly SharedAudioSystem Audio = default!;
+    [Dependency] protected readonly IGameTiming Timing = default!;
+    [Dependency] protected readonly SharedAppearanceSystem Appearance = default!;
+}
