+++ Content.Shared/Anomaly/SharedAnomalySystem.cs
@@ -352,8 +352,6 @@ public abstract class SharedAnomalySystem : EntitySystem
             }
 
             var secondsUntilNextPulse = (anomaly.NextPulseTime - Timing.CurTime).TotalSeconds;
-            UpdateScannerPulseTimers((ent, anomaly),  secondsUntilNextPulse);
-
             if (secondsUntilNextPulse < 0)
             {
                 DoAnomalyPulse(ent, anomaly);
@@ -377,22 +375,6 @@ public abstract class SharedAnomalySystem : EntitySystem
                 continue;
             DoAnomalySupercriticalEvent(ent, anom);
             // Removal of the supercritical component is handled by DoAnomalySupercriticalEvent
-        }
-    }
-
-    private void UpdateScannerPulseTimers(Entity<AnomalyComponent> anomalyEnt, double secondsUntilNextPulse)
-    {
-        if (secondsUntilNextPulse > 5)
-            return;
-        var rounded = Math.Max(0, (int)Math.Ceiling(secondsUntilNextPulse));
-
-        var scannerQuery = EntityQueryEnumerator<AnomalyScannerComponent>();
-        while (scannerQuery.MoveNext(out var scannerUid, out var scanner))
-        {
-            if (scanner.ScannedAnomaly != anomalyEnt)
-                continue;
-
-            Appearance.SetData(scannerUid, AnomalyScannerVisuals.AnomalyNextPulse, rounded);
         }
     }
 
@@ -490,6 +472,7 @@ public abstract class SharedAnomalySystem : EntitySystem
         {
             visual = AnomalyStabilityVisuals.Growing;
         }
+
         return true;
     }
